pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[project]
name = "docgen-api"
version = "0.1.0"
description = "DocGen SaaS API — FastAPI backend for documentation generation"
authors = [
    { name = "Ali Hasan", email = "ali1saleem1hasan@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic-settings",
    "python-multipart",
    "sqlalchemy",
    "alembic",
    "asyncpg",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "bcrypt==4.0.1",
    "celery",
    "redis",
    "pyyaml",
    "email-validator",
    "httpx",
    "psycopg2-binary",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
]

[tool.setuptools]
packages = {find = {where = ["."]}}

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
